+++ src/components/atoms/Button/types.d.ts
@@ -1,6 +1,12 @@
-import { CommonProps, ComponentWithChildren } from "../../../global";
+import {
+  ClickableComponent,
+  CommonProps,
+  ComponentWithChildren,
+} from "../../../global";
 
-export type ButtonProps = ComponentWithChildren &
+export type ButtonProps = ComponentWithChildren<
+  string | Element | JSX.Element
+> &
   CommonProps &
   ClickableComponent & {
     /** If true, no other styles will be applied on main container (useful for image only buttons) */