+++ src/components/molecules/RateBox/types.d.ts
@@ -1,0 +1,9 @@
+import { BoxComponent, CommonProps } from "../../../global";
+
+export type SupportedVoteType = "star" | "circle";
+
+export type RateBoxProps = CommonProps &
+  BoxComponent<number> & {
+    maxVote?: number;
+    type?: SupportedVoteType;
+  };